Refactor descriptor heap naming and stride types
Renames reservedRangeSize to reservedBytes for clarity and changes VKDescriptorRegion stride from ulong to uint for type consistency. Updates related calculations and buffer size logic to match new naming and types.

+++ sources/Zenith.NET.Vulkan/VKDescriptorRegion.cs
@@ -2,7 +2,7 @@
 
 namespace Zenith.NET.Vulkan;
 
-internal unsafe class VKDescriptorRegion(uint baseIndex, ulong stride)
+internal unsafe class VKDescriptorRegion(uint baseIndex, uint stride)
 {
     private readonly Lock @lock = new();
     private readonly Stack<uint> recycled = [];
@@ -20,7 +20,7 @@ internal unsafe class VKDescriptorRegion(uint baseIndex, ulong stride)
 
         index += baseIndex;
 
-        target = new((void*)(pointer + (nint)(stride * index)), (nuint)stride);
+        target = new((void*)(pointer + (nint)(stride * index)), stride);
 
         return new(this, index);
     }
